requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
ultralytics==8.0.200
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
scipy==1.10.1
matplotlib==3.7.2
seaborn==0.12.2
Pillow==10.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
psycopg2-binary==2.9.9